manifest: Fix CM repo URLs
Add CyanogenMod/ to the beginning of repos pulled from the "cm" remote
which does not have CyanogenMod/ in the fetch.

Rather than adding CyanogenMod/ in the "fetch" section of the remote
definition, put in in the "name" section of the projects to match how we
do this for CarbonROM repos.

Change-Id: I308f70b49a53c4ade54177f872f86a92d298ed45