Revert "manifest: track more repos from carbon"
After testing on older devices, we discovered that the VoLTE commits actually break mobile data. After research and fixing attempts, and reconsideration we came to the conclusion that breaking the majority of devices just to get VoLTE working on a very few is not worth it, at all. We should keep and eye on fixing VoLTE though.

This reverts commit a2f62b7c554eb5cb6251dacd832430a272bda9bb.

Change-Id: Ic714f2235b4eaf43f9b8c2b51f53f9b5230a7461