docs: improve manifest repo documentation
The docs included outdated build info so I tried to spruce it up a little.

Most of the changes only edit headings, add explanation, or reformat the markdown source. I also removed outdated info about using the -f parameter when syncing with repo tool as that is the default behaviour now.

Change-Id: I0c4beebe7957684972cc37892c864a796c7caa4b